test(arenabench): split the cloud CLI witnesses out of test_cloud.py
test_cloud.py is 1554 lines on main — over the 1500-line ceiling and not
grandfathered. check-file-size.sh passes today only because the ratchet
judges the change rather than the tree, so it reports the file as inherited
drift; the next PR to touch it fails. The guard's own remedy is a split
landed on its own from a fresh main, which is what this is.

The file already had a clean seam at its 'the CLI path' banner:

  cloud_fakes.py          the recorded-call AWS fakes, the executor factory
                          and the match template both halves build on
  test_cloud.py           the submit contract and the pure decisions
  test_cloud_commands.py  the _cmd_cloud_* surface: what a run refuses
                          before it spends, and fetch/status/cancel

The fakes moved to a helper module rather than being duplicated or imported
test-module-to-test-module; tests/ has no __init__.py, so pytest puts it on
sys.path and 'from cloud_fakes import ...' resolves. The naming follows the
test_cloud_merge.py / test_cloud_watch.py siblings already there.

Pure move: the 107 collected class and test names are identical before and
after (78 tests, 65 + 13), ruff check is clean, and check-file-size.sh no
longer reports drift.
