+++ arenabench/tests/test_cloud.py
@@ -14,27 +14,40 @@ billed real compute first.
 
 These tests are the witness for #2099: none of them can pass on a tree
 without ``arenabench/arenabench/cloud.py``.
+
+The ``_cmd_cloud_*`` CLI path lives in ``test_cloud_commands.py`` and the
+fakes both halves share live in ``cloud_fakes.py``.
 """
 
 from __future__ import annotations
 
 import argparse
-import io
 import json
 import tomllib
 from dataclasses import replace
-from types import SimpleNamespace
 
 import pytest
+from cloud_fakes import (
+    BUCKET,
+    GIT_URL,
+    FakeBatch,
+    FakeCodeBuild,
+    FakeDynamo,
+    FakeLsRemote,
+    FakeS3,
+    FakeSsm,
+    _executor,
+    _seat,
+    _spec,
+    _sut,
+)
 
 from arenabench.cloud import (
     BURST_QUEUE,
     JOB_DEFINITION,
     MEASURE_QUEUE,
     CloudError,
-    CloudExecutor,
     _cmd_cloud_merge,
-    _cmd_cloud_run,
     is_moving_ref,
     job_name,
     pin_slice_to_commit,
@@ -51,218 +64,9 @@ from arenabench.cloud import (
     trial_environment,
     unknown_tasks,
 )
-from arenabench.cloud_ops import _cmd_cloud_cancel, _cmd_cloud_fetch
+from arenabench.cloud_ops import _cmd_cloud_fetch
 from arenabench.config import match_from_toml
-from arenabench.model import Contestant, Engine, MatchSpec
-
-ACCOUNT = "123456789012"
-BUCKET = f"arenabench-artifacts-{ACCOUNT}"
-#: What ``infra/core.yaml``'s ``SutGitUrl`` defaults to, as the SUT build
-#: project reports it back.
-GIT_URL = "https://github.com/macanderson/stella.git"
-
-
-# --------------------------------------------------------------------------
-# recorded-call fakes — small on purpose; moto would be a heavyweight
-# dependency for what is a handful of parameter-shape assertions
-# --------------------------------------------------------------------------
-
-
-class _NoSuchKeyError(Exception):
-    pass
-
-
-class FakeS3:
-    def __init__(self, objects: dict[str, bytes] | None = None) -> None:
-        self.objects: dict[str, bytes] = dict(objects or {})
-        self.calls: list[tuple] = []
-        # boto3 spells it NoSuchKey; the adapter catches it by that name.
-        self.exceptions = SimpleNamespace(NoSuchKey=_NoSuchKeyError)
-
-    # The keyword casing is the AWS SDK's, not PEP 8's: the adapter calls
-    # with Bucket=/Key=/Body=, so the fakes must accept those spellings.
-    def put_object(self, *, Bucket: str, Key: str, Body: bytes, **_: object) -> dict:  # noqa: N803
-        self.calls.append(("put_object", Bucket, Key))
-        self.objects[Key] = Body
-        return {}
-
-    def get_object(self, *, Bucket: str, Key: str) -> dict:  # noqa: N803
-        self.calls.append(("get_object", Bucket, Key))
-        if Key not in self.objects:
-            raise _NoSuchKeyError(Key)
-        return {"Body": io.BytesIO(self.objects[Key])}
-
-    def list_objects_v2(self, **kwargs: object) -> dict:
-        self.calls.append(("list_objects_v2", kwargs.get("Prefix")))
-        prefix = str(kwargs.get("Prefix") or "")
-        keys = sorted(k for k in self.objects if k.startswith(prefix))
-        return {
-            "Contents": [{"Key": k} for k in keys],
-            "IsTruncated": False,
-        }
-
-
-class FakeBatch:
-    """Records submissions; serves job states from a per-sweep script."""
-
-    def __init__(self, state_script: list[dict[str, str]] | None = None) -> None:
-        self.submitted: list[dict] = []
-        self.describe_calls: list[list[str]] = []
-        self.state_script = list(state_script or [])
-        self._counter = 0
-
-    def submit_job(self, **kwargs: object) -> dict:
-        self.submitted.append(kwargs)
-        self._counter += 1
-        return {"jobId": f"job-{self._counter:03d}", "jobName": kwargs["jobName"]}
-
-    def describe_jobs(self, *, jobs: list[str]) -> dict:
-        self.describe_calls.append(list(jobs))
-        states = self.state_script[0] if self.state_script else {}
-        return {"jobs": [{"jobId": j, "status": states.get(j, "SUCCEEDED")} for j in jobs]}
-
-    def advance(self) -> None:
-        if len(self.state_script) > 1:
-            self.state_script.pop(0)
-
-
-class FakeSsm:
-    def __init__(self, parameters: list[str]) -> None:
-        self.parameters = parameters
-        self.calls: list[dict] = []
-
-    def get_parameters_by_path(self, **kwargs: object) -> dict:
-        self.calls.append(dict(kwargs))
-        # Two pages, to prove the paginator loop follows NextToken.
-        first, rest = self.parameters[:1], self.parameters[1:]
-        if "NextToken" not in kwargs:
-            page = {"Parameters": [{"Name": n} for n in first]}
-            if rest:
-                page["NextToken"] = "page-2"
-            return page
-        return {"Parameters": [{"Name": n} for n in rest]}
-
-
-class FakeCodeBuild:
-    def __init__(self, statuses: list[str] | None = None) -> None:
-        self.statuses = list(statuses or ["SUCCEEDED"])
-        self.started: list[dict] = []
-        self.project_calls: list[list[str]] = []
-        self.polls = 0
-
-    def start_build(self, **kwargs: object) -> dict:
-        self.started.append(dict(kwargs))
-        return {"build": {"id": "arenabench-sut-build:1234"}}
-
-    def batch_get_builds(self, *, ids: list[str]) -> dict:
-        status = self.statuses[min(self.polls, len(self.statuses) - 1)]
-        self.polls += 1
-        return {"builds": [{"id": ids[0], "buildStatus": status}]}
-
-    def batch_get_projects(self, *, names: list[str]) -> dict:
-        """The project's own ``GIT_URL`` — the remote a freshness check must
-        consult, since it is the one the buildspec clones."""
-        self.project_calls.append(list(names))
-        return {
-            "projects": [
-                {
-                    "name": names[0],
-                    "environment": {
-                        "environmentVariables": [
-                            {"name": "GIT_URL", "value": GIT_URL},
-                            {"name": "GIT_REF", "value": "main"},
-                        ]
-                    },
-                }
-            ]
-        }
-
-
-class FakeLsRemote:
-    """``git ls-remote`` as a recorded call over a fixed remote ref table."""
-
-    def __init__(self, refs: dict[str, str] | None = None) -> None:
-        self.refs = dict(refs or {})
-        self.calls: list[tuple[str, list[str]]] = []
-
-    def __call__(self, url: str, patterns) -> str:
-        patterns = list(patterns)
-        self.calls.append((url, patterns))
-        return "".join(
-            f"{self.refs[p]}\t{p}\n" for p in patterns if p in self.refs
-        )
-
-
-def _refuse_ls_remote(url: str, patterns) -> str:
-    """The seam a test uses to prove the remote is never consulted."""
-    raise AssertionError(f"git ls-remote must not run: {url} {list(patterns)}")
-
-
-class FakeSts:
-    def get_caller_identity(self) -> dict:
-        return {"Account": ACCOUNT}
-
-
-class FakeDynamo:
-    def __init__(self, pages: list[dict]) -> None:
-        self.pages = list(pages)
-        self.calls: list[dict] = []
-
-    def query(self, **kwargs: object) -> dict:
-        self.calls.append(dict(kwargs))
-        return self.pages[min(len(self.calls) - 1, len(self.pages) - 1)]
-
-
-def _executor(ls_remote=None, **clients: object) -> CloudExecutor:
-    defaults: dict[str, object] = {"sts": FakeSts()}
-    defaults.update(clients)
-    return CloudExecutor(
-        clients=defaults,
-        ls_remote=ls_remote if ls_remote is not None else _refuse_ls_remote,
-    )
-
-
-# --------------------------------------------------------------------------
-# spec helpers
-# --------------------------------------------------------------------------
-
-
-def _seat(seat_id: str, agent: str = "stella", api: str = "openrouter") -> Contestant:
-    return Contestant(
-        id=seat_id,
-        name=seat_id,
-        agent=agent,
-        engine=Engine(api=api, model="z-ai/glm-5.2"),
-        required_env=("OPENROUTER_API_KEY",),
-    )
-
-
-def _spec(
-    tasks: tuple[str, ...] = ("task-a", "task-b"),
-    seats: tuple[Contestant, ...] | None = None,
-    attempts: int = 1,
-) -> MatchSpec:
-    return MatchSpec(
-        id="m1",
-        name="cloud match",
-        dataset="terminal-bench-2.1",
-        tasks=tasks,
-        contestants=seats or (_seat("stella"), _seat("cc", agent="claude-code")),
-        attempts=attempts,
-    )
-
-
-SUT = None  # filled per-test via resolve_sut or built directly
-
-
-def _sut():
-    from arenabench.cloud import SutBinary
-
-    commit = "c" * 40
-    return SutBinary(
-        ref="main", commit=commit, uri=f"s3://{BUCKET}/binaries/main/{commit}/stella"
-    )
-
+from arenabench.model import MatchSpec
 
 # --------------------------------------------------------------------------
 # pure decisions
@@ -1149,406 +953,3 @@ class TestCloudMerge:
         rc = _cmd_cloud_merge(args)
         assert rc == 1
         assert "no results.json found" in capsys.readouterr().out
-
-
-# --------------------------------------------------------------------------
-# the CLI path
-# --------------------------------------------------------------------------
-
-TEMPLATE = """\
-[match]
-name = "cloud smoke"
-dataset = "terminal-bench-2.1"
-tasks = ["task-a"]
-
-[[contestant]]
-id = "stella"
-agent = "stella"
-
-  [contestant.engine]
-  api = "openrouter"
-  model = "z-ai/glm-5.2"
-
-  [contestant.env]
-  required = ["OPENROUTER_API_KEY"]
-"""
-
-
-def _run_args(template, **overrides) -> argparse.Namespace:
-    defaults = dict(
-        template=str(template),
-        ref=None,
-        run_id="r-test",
-        burst=False,
-        poll=0.0,
-        vcpus=4,
-        memory_mb=15360,
-        no_wait=True,
-        no_fetch=True,
-        artifacts=False,
-        out=None,
-        region=None,
-        bucket=None,
-        # These tests are about submission, not about the banned-behavior
-        # gate, so they take the explicit waiver — `_cmd_cloud_run` refuses to
-        # start a run that is neither gated nor deliberately ungated, and that
-        # refusal has its own witness in tests/test_gate.py.
-        gate=None,
-        no_gate=True,
-        gate_allow=None,
-    )
-    defaults.update(overrides)
-    return argparse.Namespace(**defaults)
-
-
-class TestCloudRunCommand:
-    @pytest.fixture(autouse=True)
-    def _no_local_dataset(self, monkeypatch):
-        """Task-id validation (#3208) reads the machine's dataset registry.
-
-        Pinned to an empty registry so these tests behave identically on a
-        machine with terminal-bench materialised and on CI, where it is not:
-        an unfetched dataset is a declared validation gap that submits, and
-        that is the neutral posture every non-#3208 test here wants.
-        """
-        import arenabench.registry as registry_mod
-
-        monkeypatch.setattr(
-            registry_mod, "DEFAULT_REGISTRY", registry_mod.Registry()
-        )
-
-    def test_a_seat_without_ssm_credentials_is_refused_before_any_submit(
-        self, tmp_path, capsys
-    ) -> None:
-        """The refusal half of the witness: nothing is uploaded and no job is
-        submitted when a seat's declared credentials will not exist in the
-        trial environment (#1827)."""
-        template = tmp_path / "match.toml"
-        template.write_text(TEMPLATE, encoding="utf-8")
-        s3, batch = FakeS3(), FakeBatch()
-        executor = _executor(s3=s3, batch=batch, ssm=FakeSsm([]))
-
-        rc = _cmd_cloud_run(
-            _run_args(template),
-            executor=executor,
-            task_lister=lambda _dataset: ["task-a"],
-        )
-
-        assert rc == 2
-        assert batch.submitted == [], "a refused match must submit nothing"
-        assert s3.calls == [], "a refused match must upload nothing"
-        err = capsys.readouterr().err
-        assert "OPENROUTER_API_KEY" in err
-        assert "refusing to submit" in err
-
-    def test_a_task_the_dataset_does_not_contain_is_refused_before_any_submit(
-        self, tmp_path, capsys
-    ) -> None:
-        """The refusal half of #3255's witness: gate89high1, fulleightynine1
-        and h2h891 each submitted a spec whose first sorted task was the
-        literal count "89" (a task-listing capture's trailer), and each
-        burned its index-000 container on it. With the dataset enumerable
-        locally, the submission is refused before any upload or job."""
-        template = tmp_path / "match.toml"
-        template.write_text(
-            TEMPLATE.replace('tasks = ["task-a"]', 'tasks = ["89", "task-a"]'),
-            encoding="utf-8",
-        )
-        s3, batch = FakeS3(), FakeBatch()
-        executor = _executor(s3=s3, batch=batch, ssm=FakeSsm([]))
-
-        rc = _cmd_cloud_run(
-            _run_args(template),
-            executor=executor,
-            task_lister=lambda _dataset: ["task-a", "task-b"],
-        )
-
-        assert rc == 2
-        assert batch.submitted == [], "a refused match must submit nothing"
-        assert s3.calls == [], "a refused match must upload nothing"
-        err = capsys.readouterr().err
-        assert "'89'" in err
-        assert "does not contain" in err
-        assert "task-a" not in err, "only the bogus names are refused"
-
-    def test_an_unmaterialised_dataset_submits_but_says_it_was_not_checked(
-        self, tmp_path, capsys, monkeypatch
-    ) -> None:
-        """The declared gap: with no local task tree to check against, the
-        submission proceeds — refusing would make local materialisation a
-        submit prerequisite — but the skipped validation is said out loud."""
-        monkeypatch.delenv("ARENABENCH_STELLA_REPO", raising=False)
-        commit = "d" * 40
-        template = tmp_path / "match.toml"
-        template.write_text(TEMPLATE, encoding="utf-8")
-        s3 = FakeS3(
-            {
-                "binaries/main/latest.json": json.dumps(
-                    {"git_ref": "main", "commit": commit}
-                ).encode()
-            }
-        )
-        batch = FakeBatch()
-        executor = _executor(
-            s3=s3,
-            batch=batch,
-            ssm=FakeSsm(["/arenabench/openrouter_api_key"]),
-            codebuild=FakeCodeBuild(),
-            ls_remote=FakeLsRemote({"refs/heads/main": commit}),
-        )
-
-        rc = _cmd_cloud_run(
-            _run_args(template),
-            executor=executor,
-            task_lister=lambda _dataset: [],
-        )
-
-        assert rc == 0
-        assert len(batch.submitted) == 1
-        out = capsys.readouterr().out
-        assert "task names not validated" in out
-
-    def test_a_credentialled_match_submits_with_the_matchs_own_ref(
-        self, tmp_path, capsys, monkeypatch
-    ) -> None:
-        """End to end through the CLI glue: SSM credentials present, the SUT
-        resolved from the match's default ref (main), one job per trial."""
-        monkeypatch.delenv("ARENABENCH_STELLA_REPO", raising=False)
-        commit = "d" * 40
-        template = tmp_path / "match.toml"
-        template.write_text(TEMPLATE, encoding="utf-8")
-        s3 = FakeS3(
-            {
-                "binaries/main/latest.json": json.dumps(
-                    {"git_ref": "main", "commit": commit}
-                ).encode()
-            }
-        )
-        batch = FakeBatch()
-        executor = _executor(
-            s3=s3,
-            batch=batch,
-            ssm=FakeSsm(["/arenabench/openrouter_api_key"]),
-            codebuild=FakeCodeBuild(),
-            ls_remote=FakeLsRemote({"refs/heads/main": commit}),
-        )
-
-        rc = _cmd_cloud_run(
-            _run_args(template),
-            executor=executor,
-            task_lister=lambda _dataset: ["task-a"],
-        )
-
-        assert rc == 0
-        (submitted,) = batch.submitted
-        env = {
-            e["name"]: e["value"]
-            for e in submitted["containerOverrides"]["environment"]
-        }
-        assert env["SUT_COMMIT"] == commit
-        assert env["RUN_ID"] == "r-test"
-        assert "runs/r-test/jobs.json" in s3.objects
-        out = capsys.readouterr().out
-        assert "cloud status r-test" in out
-        assert f"sut       : main -> {commit}" in out, (
-            "submit-time output must name ref -> commit, so a reader can see "
-            "which commit 'main' meant on this run (#2388)"
-        )
-
-
-class TestSubmitTaskValidation:
-    """#3208: a task id the dataset does not contain must not become a job.
-
-    A phantom "89" pasted over `write-compressor` burned two Batch containers
-    and silently shrank the "full 89" panel to 88 tasks.
-    """
-
-    def _registry_with(self, monkeypatch, names: list[str]) -> None:
-        import arenabench.registry as registry_mod
-
-        fake = SimpleNamespace(
-            tasks=lambda key: [SimpleNamespace(name=n) for n in names]
-        )
-        monkeypatch.setattr(registry_mod, "DEFAULT_REGISTRY", fake)
-
-    def test_an_unknown_task_id_refuses_before_any_upload_or_job(
-        self, tmp_path, capsys, monkeypatch
-    ) -> None:
-        self._registry_with(monkeypatch, ["task-b", "write-compressor"])
-        template = tmp_path / "match.toml"
-        template.write_text(TEMPLATE, encoding="utf-8")
-        s3, batch = FakeS3(), FakeBatch()
-        executor = _executor(s3=s3, batch=batch, ssm=FakeSsm(["openrouter_api_key"]))
-
-        rc = _cmd_cloud_run(_run_args(template), executor=executor)
-
-        assert rc == 2
-        assert batch.submitted == [], "a refused match must submit nothing"
-        assert s3.calls == [], "a refused match must upload nothing"
-        err = capsys.readouterr().err
-        assert "task-a" in err, "the refusal must name the offending id"
-        assert "refusing to submit" in err
-
-    def test_a_known_task_list_submits_and_an_unfetched_dataset_says_so(
-        self, tmp_path, capsys, monkeypatch
-    ) -> None:
-        """An empty registry is a validation gap, not a refusal — and the gap
-        is said out loud rather than silently waved through."""
-        self._registry_with(monkeypatch, [])
-        template = tmp_path / "match.toml"
-        template.write_text(TEMPLATE, encoding="utf-8")
-        executor = _executor(
-            s3=FakeS3(), batch=FakeBatch(), ssm=FakeSsm(["openrouter_api_key"])
-        )
-
-        # The run stops later, at the missing --ref — which is the point:
-        # validation let it through and said why it could not check.
-        rc = _cmd_cloud_run(_run_args(template, ref=""), executor=executor)
-
-        captured = capsys.readouterr()
-        assert "not validated" in captured.out
-        assert rc == 2 and "Pass --ref" in captured.err
-
-
-class TestCloudFetchGate:
-    """#3218: a fetch of a live run is a silently short match without --partial."""
-
-    def _executor_with_states(self, states: dict[str, str]):
-        s3 = FakeS3({})
-        executor = _executor(
-            s3=s3, batch=FakeBatch([states]), dynamodb=FakeDynamo([{"Items": []}])
-        )
-        executor.load_jobs = lambda run_id: {  # type: ignore[method-assign]
-            "run_id": run_id,
-            "queue": "arenabench-measure",
-            "jobs": [
-                {"id": "job-1", "label": "000-stella-task-a", "task": "task-a"},
-                {"id": "job-2", "label": "001-stella-task-b", "task": "task-b"},
-            ],
-        }
-        return executor
-
-    def test_a_live_run_is_refused_without_partial(self, tmp_path, capsys) -> None:
-        executor = self._executor_with_states(
-            {"job-1": "SUCCEEDED", "job-2": "RUNNING"}
-        )
-        args = argparse.Namespace(
-            run_id="r1", artifacts=False, out=str(tmp_path), region=None,
-            bucket=None, partial=False,
-        )
-        rc = _cmd_cloud_fetch(args, executor=executor)
-        assert rc == 2
-        err = capsys.readouterr().err
-        assert "001-stella-task-b" in err and "RUNNING" in err
-        assert "--partial" in err
-
-    def test_partial_fetches_and_records_the_states_it_saw(self, tmp_path) -> None:
-        executor = self._executor_with_states(
-            {"job-1": "SUCCEEDED", "job-2": "RUNNING"}
-        )
-        args = argparse.Namespace(
-            run_id="r1", artifacts=False, out=str(tmp_path), region=None,
-            bucket=None, partial=True,
-        )
-        rc = _cmd_cloud_fetch(args, executor=executor)
-        assert rc == 0
-        sidecar = json.loads((tmp_path / "jobs.json").read_text())
-        by_id = {j["id"]: j["state"] for j in sidecar["jobs"]}
-        assert by_id == {"job-1": "SUCCEEDED", "job-2": "RUNNING"}
-
-    def test_a_settled_run_fetches_without_ceremony(self, tmp_path) -> None:
-        executor = self._executor_with_states(
-            {"job-1": "SUCCEEDED", "job-2": "FAILED"}
-        )
-        args = argparse.Namespace(
-            run_id="r1", artifacts=False, out=str(tmp_path), region=None,
-            bucket=None, partial=False,
-        )
-        assert _cmd_cloud_fetch(args, executor=executor) == 0
-
-
-class TestStatusReconciliation:
-    """#3217: a ddb row an externally-killed trial never wrote must not read
-    `running` when Batch says the job is terminal."""
-
-    def test_the_killed_trial_shape_renders_terminal(self) -> None:
-        from arenabench.cloud_ops import reconciled_status
-
-        assert (
-            reconciled_status("running", "FAILED") == "failed (killed; no checkin)"
-        )
-        assert (
-            reconciled_status("running", "SUCCEEDED")
-            == "succeeded (no terminal checkin)"
-        )
-
-    def test_everything_else_prints_the_row_as_written(self) -> None:
-        from arenabench.cloud_ops import reconciled_status
-
-        assert reconciled_status("running", "RUNNING") == "running"
-        assert reconciled_status("running", None) == "running"
-        assert reconciled_status("finished", "FAILED") == "finished"
-
-
-class TestCloudCancelCommand:
-    """#3257: `cloud cancel` reaches cancel_run — the safe, jobs.json-scoped
-    stop — instead of leaving operators to hand-roll a name-prefix sweep."""
-
-    def _args(self, run_id: str = "r1", reason: str | None = None):
-        return argparse.Namespace(
-            run_id=run_id, reason=reason, region=None, bucket=None
-        )
-
-    def test_a_run_with_no_live_jobs_is_a_clean_no_op(self, capsys) -> None:
-        batch = FakeBatch([{"job-1": "SUCCEEDED", "job-2": "FAILED"}])
-        executor = _executor(batch=batch)
-        executor.load_jobs = lambda run_id: {  # type: ignore[method-assign]
-            "jobs": [{"id": "job-1"}, {"id": "job-2"}]
-        }
-        rc = _cmd_cloud_cancel(self._args(), executor=executor)
-        assert rc == 0
-        out = capsys.readouterr().out
-        assert "2 already settled" in out
-        assert "0 terminated" in out and "0 cancelled" in out
-
-    def test_live_jobs_are_signalled_and_the_split_is_printed(self, capsys) -> None:
-        class _CancellingBatch(FakeBatch):
-            def __init__(self, script):
-                super().__init__(script)
-                self.cancelled: list[str] = []
-                self.terminated: list[str] = []
-
-            def describe_jobs(self, *, jobs):
-                # The real API drops ids it cannot resolve rather than
-                # erroring; the base fake fabricates SUCCEEDED instead, which
-                # is exactly the shape this test must not inherit.
-                states = self.state_script[0] if self.state_script else {}
-                return {
-                    "jobs": [
-                        {"jobId": j, "status": states[j]}
-                        for j in jobs
-                        if j in states
-                    ]
-                }
-
-            def cancel_job(self, *, jobId, reason):  # noqa: N803 - Batch's spelling
-                self.cancelled.append(jobId)
-
-            def terminate_job(self, *, jobId, reason):  # noqa: N803 - Batch's spelling
-                self.terminated.append(jobId)
-
-        batch = _CancellingBatch([{"job-1": "RUNNING", "job-2": "RUNNABLE"}])
-        executor = _executor(batch=batch)
-        executor.load_jobs = lambda run_id: {  # type: ignore[method-assign]
-            # job-3 is absent from describe_jobs: still cancelled, never
-            # skipped — an id we own that a describe cannot resolve may still
-            # be billing (the branch cancel.plan_cancellation takes on
-            # purpose).
-            "jobs": [{"id": "job-1"}, {"id": "job-2"}, {"id": "job-3"}]
-        }
-        rc = _cmd_cloud_cancel(self._args(reason="smoke"), executor=executor)
-        assert rc == 0
-        assert batch.terminated == ["job-1"]
-        assert set(batch.cancelled) == {"job-2", "job-3"}
-        out = capsys.readouterr().out
-        assert "1 terminated" in out and "2 cancelled" in out
